make service init.py clean by creating app.py in service/src. change imports in test_routes, integration test, e2e test.
