+++ tests/integration/test_service_adapter_integration.py
@@ -12,7 +12,7 @@ from starlette.types import ASGIApp
 
 import gmail_client_impl
 import mail_client_api
-import mail_client_service
+from mail_client_service.app import app, get_mail_client, _client_factory
 import mail_client_adapter
 from mail_client_adapter import ServiceMailClient
 
@@ -61,9 +61,9 @@ def test_mail_client_adapter_round_trip_through_service() -> None:
     gmail_mock.mark_as_read.return_value = True
     gmail_mock.delete_message.return_value = True
 
-    app = mail_client_service.app
-    app.dependency_overrides[mail_client_service.get_mail_client] = lambda: gmail_mock
-    mail_client_service._client_factory.cache_clear()
+    # app = mail_client_service.app
+    app.dependency_overrides[get_mail_client] = lambda: gmail_mock
+    _client_factory.cache_clear()
 
     base_url = "http://testserver"
     transport = _build_sync_transport(app)
@@ -97,6 +97,6 @@ def test_mail_client_adapter_round_trip_through_service() -> None:
         gmail_mock.delete_message.assert_called_once_with(sample_message.id)
     finally:
         transport.close()
-        app.dependency_overrides.pop(mail_client_service.get_mail_client, None)
-        mail_client_service._client_factory.cache_clear()
+        app.dependency_overrides.pop(get_mail_client, None)
+        _client_factory.cache_clear()
         gmail_client_impl.register()
